replace custom test collector and runner with pytest and tox  (#39)
* check cache directory exists before emptying

When called via Compiler destructor directory may already have been
removed. Seen during testing.

* replace custom test collector and runner with pytest and tox

Removes a significant amount of code and enables easy testing of all
supported versions of Python.

Hide deprecation warnings for known outstanding GTK issues.

+++ fract4d/cache.py
@@ -32,6 +32,8 @@ class T:
         self.createPickledFile(self._indexName(), self.files)
     
     def clear(self):
+        if not os.path.exists(self.dir):
+            return
         for f in os.listdir(self.dir):
             try:
                 os.remove(os.path.join(self.dir,f))
